Move `UnsubscribeReason` to the `Subscriptions` endpoint types
