--- src/types/UnsubscribeReason.ts
@@ -1,7 +1,0 @@
-export enum UnsubscribeReason {
-    UNKNOWN = 'unknown',
-    UNWANTED = 'unwanted',
-    SPAM = 'spam',
-    NEVER_SIGNED_UP = 'never_signed_up',
-    OTHER = 'other',
-}
