+++ src/types/index.ts
@@ -34,7 +34,6 @@ export * from './SenderDomain';
 export * from './Story';
 export * from './Snippet';
 export * from './UserRef';
-export * from './UnsubscribeReason';
 export * from './EmailBranding';
 export * from './EmailBrandingMode';
 export * from './PrivacyRequest';
